tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/size.attributes.parse.onlyspace.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: size.attributes.parse.onlyspace</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>size.attributes.parse.onlyspace</h1>
<p class="desc">Parsing of non-negative integers</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="  " height="  "><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="size.attributes.parse.onlyspace.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Parsing of non-negative integers");
_addTest(function(canvas, ctx) {

_assertSame(canvas.width, 300, "canvas.width", "300");
_assertSame(canvas.height, 150, "canvas.height", "150");
_assertSame(canvas.getAttribute('width'), '  ', "canvas.getAttribute('width')", "'  '");
_assertSame(canvas.getAttribute('height'), '  ', "canvas.getAttribute('height')", "'  '");


});
</script>
